requirement.txt
google-generativeai
python-dotenv
streamlit
PyPDF2
python-docx
